legacy/diffused-rays/requirements.txt
# Core dependencies
pygame>=2.5.0
numpy>=1.24.0
Pillow>=10.0.0

# PyTorch with MPS support (Apple Silicon)
torch>=2.1.0

# Stable Diffusion (for later milestones)
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
